+++ src/pyensae/pandas_helper/__init__.py
@@ -5,3 +5,4 @@
 
 from .joins import df_crossjoin
 from .missing import add_missing_indices
+from .table_formula import TableFormula