+++ HISTORY.rst
@@ -11,3 +11,4 @@ History
 * `9`: switch from Yahoo Finance to Google as the default provider
   for financial data
 * `11`: create graph image with blockdiag
+* `13`: add a class *TableFormula* as an example to extend pandas dataframe